refactor(app): move the record-only write path into its own file
WatchLoop.swift sat at 599 lines against a 600-line --strict cap, so any
further change to it trips file_length and type_body_length. Move the
record-only branch (writeRecordOnlySidecar + its move helper) into
WatchLoop+RecordOnly.swift, matching the existing line-cap splits
(PipelineQueue+Stages, PipelineQueue+Recovery).

The moved method now throws instead of handling its own failure. Reporting
a failure means writing lastError, and that goes through the private
`update` funnel, which an extension in another file cannot reach. Letting
the write path throw and catching at its single caller keeps every
observable write behind the funnel; the alternatives were widening `update`
to module scope or parking a record-only-specific reporting method next to
it, and both erode the funnel's single-entry property.

Record-only was the only sizeable block that touches `update` just once,
which is what made it the viable cut. No behavior change.
